ensure that only a single copy of Solid is used
This stores Solid APIs on global variables, then if any duplicate lume/variable package is ever loaded, it will use the global version of Solid. This will be ok except when trying to use multiple incompatible (non-semver compatible) versions of lume packages in the same app, which should be avoided.